fix(week-C-stubs): close all R3 findings — HIGH-1/2, MED-1/3, LOW-1/3
HIGH-1: app/request/page.tsx — <Link href="mailto:..."> → <a href="mailto:...">
  (Next.js Link is for internal routes; external/mailto must use native <a>)

HIGH-2: app/__tests__/request-page.test.tsx — add contract test suite
  (heading render + CTA mailto href + back-link /dashboard assertions)

MED-1: components/ui/card.tsx — CardTitle div → h3
  (semantic heading hierarchy; used by upgrade page tier names under h1)

MED-3: delete __tests__/pages/upgrade-page.test.tsx + matches-page.test.tsx
  (legacy placeholder suites superseded by contract versions in app/__tests__)

LOW-1: app/matches/page.tsx — key={idx} → key={vessel-date} stable composite
  (idx keys cause silent re-order bugs; vessel+date is stable per entry)

LOW-3: app/upgrade/page.tsx — TIERS: Tier[] explicit type annotation
  (prevents silent structural drift on future TIERS array edits)

Tests: 126/126 pass
